Fix long-op review findings: opId correlation, cancel ownership, deadlines
- Decode failures extract opId via OpRequestShell so clients get taxonomy errors
- Cancel and worker share write ownership through atomic inFlight remove
- Client waits with grace after deadline so server timeout can arrive
- Server schedules mid-op deadline interrupt as taxonomy timeout
- Extract MultiplexedIpcSession for detekt TooManyFunctions
- Widen SpectreProcesses list drift tolerance for busy hosts
